Refactored recepteur, renamed measuresrtrv to measuresdata, passed redis funcs to measuresdata
