+++ internal/measuresdata/measuresdata.go
@@ -1,12 +1,21 @@
-package measuresrtrv
+package measuresdata
 
 import (
 	data "Moniport/internal/data"
 	redis "Moniport/internal/helpers/redis"
+	"fmt"
 	"strconv"
 	"strings"
 	"time"
 )
+
+func Connect() {
+	redis.Connect()
+}
+
+func Disconnect() {
+	redis.CloseConnection()
+}
 
 func GetAirports() []string {
 	return redis.GetSet("airports")
@@ -36,4 +45,38 @@ func parseMeasures(airport string, measureType string, measures map[int64]string
 		parsedMeasures = append(parsedMeasures, measure)
 	}
 	return parsedMeasures
+}
+
+func SendMeasure(m data.Measure) {
+	setKey := m.IDAirport + ":" + m.MeasureType
+
+	redis.AddToSet("airports", m.IDAirport)
+
+	setValue := fmt.Sprintf("%d_%.2f", getNewIdMeasure(), m.Value)
+
+	setTimestamp := getTimestampFromDate(m.Date)
+
+	redis.AddToOrdSet(setKey, setValue, setTimestamp)
+}
+
+func getNewIdMeasure() int {
+	if redis.KeyExists("currIdMeasure") {
+		redis.IncrKey("currIdMeasure")
+	} else {
+		redis.SendData("currIdMeasure", "0")
+	}
+
+	return redis.GetDataInt("currIdMeasure")
+}
+
+func getTimestampFromDate(date string) int64 {
+	layout := "2006-01-02-15-04-05"
+
+	t, err := time.Parse(layout, date)
+
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	return t.Unix()
 }
